deleted layout, added Prisma generic type
