+++ components/newPoll/CreatePoll.tsx
@@ -4,11 +4,11 @@ import InputField from '../InputField';
 export default function CreatePoll() {
   const { register } = useFormContext(); // retrieve all hook methods
   return (
-    <>
-      <div>Deadline</div>
+    <div className="flex flex-col gap-4">
+      <h3>Question & Details</h3>
       <InputField
         required
-        {...register('test3')}
+        {...register('question')}
         label={'Create a poll'}
         showLabel={false}
         type={'text'}
@@ -17,13 +17,22 @@ export default function CreatePoll() {
       />
       <InputField
         required
-        {...register('test3')}
+        {...register('description')}
         label={'Create a poll'}
         showLabel={false}
         type={'text'}
         width={'full'}
         disabled={false}
       />
-    </>
+      <InputField
+        required
+        {...register('random name')}
+        label={'Create a poll'}
+        showLabel={false}
+        type={'text'}
+        width={'full'}
+        disabled={false}
+      />
+    </div>
   );
 }
